* ARTTARIFLIGNE.cs: working on buy
* TIERS.cs:
* PIECEACHAT.cs:
* PIECEVENTE.cs:
* DataServices.cs:
* RestServices.cs:
* REGLEMENTECHEANCE.cs:
* PIECEVENTEECHEANCE.cs:
* PIECEACHATECHEANCE.cs:
* BuyFRentPageModel.cs:
* BuyFAentPageModel.cs:
* BuyBCLigPageModel.cs:
* BuyBCEntPageModel.cs:
* SellBREntPageModel.cs:
* SellFREntPageModel.cs:
* SellFVEntPageModel.cs:
* SellLignePageModel.cs:
* SellBLEntPageModel.cs:
* BuyDetailsPageModel.cs:
* SellEntetePageModel.cs:
* StockLignePageModel.cs:
* SellDetailsPageModel.cs:
* BuyBRecentetPageModel.cs:
* BuyBRetEntetPageModel.cs:

+++ PFE/PFE/PageModels/BuyDetailsPageModel.cs
@@ -36,19 +36,31 @@ namespace PFE.PageModels
                     isBusy = true;
                 });
 
-                if (await _restservices.PostBuyElements(productList))
+                foreach (Buyelement s in productList)
                 {
-
-                    if (_dataService.RemoveBuyElements())
+                    if (!s.ligneUpdated)
                     {
-                        productList.Clear();
-                        _dialogService.ShowMessage("success", false);
+                        productList.Remove(s);
                     }
-                    else
+                }
+                if(productList.Count>0){
+                    if (await _restservices.PostBuyElements(productList))
                     {
+
+                        if (_dataService.RemoveBuyElements())
+                        {
+                            productList.Clear();
+                            _dialogService.ShowMessage("success", false);
+                        }
+                        else
+                        {
+                            _dialogService.ShowMessage("erreur , veuillez reessayer plus tard", true);
+                        }
+                    }else{
                         _dialogService.ShowMessage("erreur , veuillez reessayer plus tard", true);
                     }
                 }
+
 
                 Device.BeginInvokeOnMainThread(() =>
                 {
